autotest_package/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]
python-dotenv==1.0.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0

# RabbitMQ dependencies
celery==5.3.4
kombu==5.3.4
pika==1.3.2
aio-pika==9.3.1
requests==2.31.0
